MdePkg ACPI: Correct processor struct of PPTT
The Type field of EFI_ACPI_6_2_PPTT_STRUCTURE_PROCESSOR should
be UINT8 as ACPI version 6.2 specification.

Contributed-under: TianoCore Contribution Agreement 1.1
(cherry picked from commit 19ef86eec6ed67c11b6a942865b143123180d394)

+++ MdePkg/Include/IndustryStandard/Acpi62.h
@@ -2569,7 +2569,7 @@ typedef struct {
 /// Processor hierarchy node structure
 ///
 typedef struct {
-  UINT32                                        Type;
+  UINT8                                         Type;
   UINT8                                         Length;
   UINT8                                         Reserved[2];
   EFI_ACPI_6_2_PPTT_STRUCTURE_PROCESSOR_FLAGS   Flags;
